+++ internal/toolcall/file/file.go
@@ -24,9 +24,9 @@ func ResolvePath(ctx context.Context, path string) string {
 // ParseLineRange parse line range
 func ParseLineRange(args toolcall.ToolArgs) (int, int, error) {
 	// 解析开始行号
-	startLine := toolcall.ToolArgsValue(args, "start_line", 1)
+	startLine := int(toolcall.ToolArgsValue(args, "start_line", int64(1)))
 	// 解析结束行号
-	endLine := toolcall.ToolArgsValue(args, "end_line", -1) // -1 表示到文件末尾
+	endLine := int(toolcall.ToolArgsValue(args, "end_line", int64(-1))) // -1 表示到文件末尾
 	// 验证行号范围
 	if endLine != -1 && endLine < startLine {
 		return 0, 0, fmt.Errorf("end_line must be greater than or equal to start_line")
