+++ internal/toolcall/issue/issue_show.go
@@ -10,7 +10,7 @@ import (
 
 // handleIssueShow 处理显示单个issue（Tool Calling）
 func handleIssueShow(ctx context.Context, args toolcall.ToolArgs) (output string, user string, err error) {
-	number := toolcall.ToolArgsValue(args, "number", 0)
+	number := int(toolcall.ToolArgsValue(args, "number", int64(0)))
 	if number == 0 {
 		err = fmt.Errorf("必须提供issue编号")
 		return
